admin.html 을 fetch_admin.html 로 변경하였습니다 로그인페이지에서 다른페이지로 넘어갈때, 로그인 세선기능으로 인증하여 넘어가도록 수정하였습니다. 리뷰를 삭제기능을 완성하였습니다. 리뷰삭제 성공,실패, 이미지 디렉토리 삭제 성공,실패,없음 상태코드도 추가되었습니다.

+++ client_data/client_review_delete.php
@@ -13,15 +13,35 @@ if(!$conn = connect_mysqli(MYSQL_IP,MAIN_DB,DB_PASSWORD,USE_DB)){
 else{
 //echo "connect success!\n";
 }
+$sql = "SELECT pid, image_num FROM ".REVIEW_TABLE."{$value['user_id']} WHERE pid = {$value['store_id']}";
+if(!$result = mysqli_query($conn,$sql)){
+	$to_client['state'] = SQL_QUERY_ERROR;
+}else{
+	$row = mysqli_fetch_assoc($result);
+	$image_num = $row['image_num'];
+}
 
-$sql = "DELETE FROM ".REVIEW_TABLE." WHERE pid = {$value['store_id']}";
+$sql = "DELETE FROM ".REVIEW_TABLE."{$value['user_id']} WHERE pid = {$value['store_id']}";
 if(!mysqli_query($conn,$sql)){
 		// insert fail
-	$to_client['state'] = $sql;
+	$to_client['state'] = CLIENT_REVIEW_DATA_DELETE_FAIL;
 }else{
-	$to_client['state'] = "review delete complete";
-	$target_dir_name = "./client_{$value['user_id']}_{$value['map_id']}_{$value['store_id']}"; 
-	$to_client['rmdir_state'] = removeDirectory($target_dir_name);
+	$to_client['state'] = CLIENT_REVIEW_DATA_DELETE_SUCCESS;
+	
+	if($image_num!=0){
+		$target_dir_name = "./client_{$value['user_id']}_{$value['map_id']}_{$value['store_id']}";
+		if(removeDirectory($target_dir_name)){	
+			$to_client['rmdir_state'] = CLIENT_REVIEW_IMAGE_RMDIR_SUCCESS;
+			//$to_client['rmdir_state'] = $temp;
+		}else{
+			$to_client['rmdir_state'] = CLIENT_REVIEW_IMAGE_RMDIR_FAIL;
+			//$to_client['rmdir_state'] = $temp;
+		}
+	}else{			
+		$to_client['rmdir_state'] = CLIENT_REVIEW_IMAGE_RMDIR_NONE;
+	}
 }
+	
+
 echo json_encode($to_client);
 ?>
